feat: add upload user photo endpoint

+++ src/services/user-service.js
@@ -427,7 +427,7 @@ export default class UserService {
                 );
             });
 
-            if (storeResponse?.path) {
+            if (storeResponse?.path && user.user_profile.photo) {
                 await this.storage.delete(user.user_profile.photo.replace(ASSET_URL, S3_OBJECT_URL), { s3: { bucket: process.env.S3_BUCKET_NAME } });
             }
